+++ src/main/java/rentalstore/NewRelease.java
@@ -6,7 +6,7 @@ public class NewRelease extends MovieType{
         return  thisAmount;
     }
 
-    public int updateFrequentRenterPointers(int frequentRenterPoints, int dayRented){
+    public double updateFrequentRenterPointers(double frequentRenterPoints, int dayRented){
         if (dayRented > 1) {
             frequentRenterPoints++;
         }
